requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appnope==0.1.0            # via ipython
backcall==0.1.0           # via ipython
cached-property==1.4.2    # via docker-compose
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
click==6.7                # via rq
decorator==4.3.0          # via ipython, traitlets
django-cors-headers==2.2.0
django-environ==0.4.4
django-extensions==2.0.7
django-filter==1.1.0
django-rq==1.1.0
django==2.0.5
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2
docker-compose==1.21.2
docker-pycreds==0.2.3     # via docker
docker==3.3.0             # via docker-compose
dockerpty==0.4.1          # via docker-compose
docopt==0.6.2             # via docker-compose
gevent==1.3.0
greenlet==0.4.13          # via gevent
gunicorn==19.8.1
idna==2.6                 # via requests
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==6.4.0
jedi==0.12.0              # via ipython
jsonschema==2.6.0         # via docker-compose
model-mommy==1.5.1
parso==0.2.0              # via jedi
pexpect==4.5.0            # via ipython
pickleshare==0.7.4        # via ipython
prompt-toolkit==1.0.15    # via ipython
psycopg2-binary==2.7.4
ptyprocess==0.5.2         # via pexpect
pygments==2.2.0           # via ipython
pyjwt==1.6.1              # via djangorestframework-jwt
pytz==2018.4              # via django
pyyaml==3.12              # via docker-compose
redis==2.10.6             # via rq
requests==2.18.4
rq==0.10.0                # via django-rq
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via django-environ, django-extensions, docker, docker-compose, docker-pycreds, dockerpty, model-mommy, prompt-toolkit, traitlets, websocket-client
texttable==0.9.1          # via docker-compose
traitlets==4.3.2          # via ipython
urllib3==1.22             # via requests
wcwidth==0.1.7            # via prompt-toolkit
websocket-client==0.47.0  # via docker, docker-compose
werkzeug==0.14.1
